% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enercast.R
\docType{package}
\name{enercast}
\alias{enercast}
\alias{enercast-package}
\title{enercast: A package for predicting electricity load.}
\description{
The enercast package provides different models to predict the daily electricity load. Models covers
both endogenous and exogenous variants (e.g. using or not external information as for instance
meteorological variables).
}
\section{enercast models}{

hong, ssm, ...
}

