% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{performance}
\alias{performance}
\title{performance}
\usage{
performance(y, yhat, measure = "rmse")
}
\arguments{
\item{y}{}

\item{yhat}{}

\item{measure}{}
}
\value{
single value
}
\description{
Performance measure for a predicted set of values.
}
\examples{
performance(rep(0, 10), rnorm(10))
}
