% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.ssm.R
\name{pred.ssm}
\alias{pred.ssm}
\title{predict.ssm}
\usage{
pred.ssm(ssm.obj.i)
}
\arguments{
\item{ssm.obj}{Estimated object from \code{ssm}}

\item{newdata}{data}

\item{for.each}{'hour'}
}
\value{
list
}
\description{
Prediction in Univariate State Space Model
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
