% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rf.R
\name{predict.rf}
\alias{predict.rf}
\title{predict.rf}
\usage{
\method{predict}{rf}(listRF, newdata = NULL, for.each = "hour")
}
\arguments{
\item{listRF}{List of Random Forest Models}

\item{newdata}{New data to predict}

\item{for.each}{Character. Variable name of data frequency (i.e. 'hour')}
}
\value{
list
}
\description{
Prediction in Random Forest
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
