% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sparse.R
\name{predict.sparse}
\alias{predict.sparse}
\title{predict.sparse}
\usage{
\method{predict}{sparse}(sp, new.data, delta = 24)
}
\arguments{
\item{sp}{}

\item{new.data}{}

\item{delta}{}
}
\description{
Compute prediction for sparse model
}
