% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ssm.R
\name{predict.ssm}
\alias{predict.ssm}
\title{Title}
\usage{
\method{predict}{ssm}(ssm.obj, newdata, for.each = "hour")
}
\arguments{
\item{ssm.obj}{}

\item{newdata}{}

\item{for.each}{}
}
\description{
Title
}
