% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.wavkerfun.R
\name{predict.wavkerfun}
\alias{predict.wavkerfun}
\title{Functional wavelet kernal}
\usage{
\method{predict}{wavkerfun}(object, ..., newdata = NULL, sd)
}
\arguments{
\item{...}{}

\item{newdata}{}

\item{sd}{}

\item{wavkerfun}{object}
}
\value{
predicted values
}
\description{
Functional wavelet kernal
}
