\name{rcs}
\alias{rcs}
\title{Construct splines}
\description{algo}
\usage{rcs(x, m = 5, nod = NULL)}

\arguments{
  \item{x}{numeric vector}
  \item{m}{knots number (default=5)}
  \item{nod}{knots placement}
}
\value{matrix with normalized splines}

\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}


