\name{rf}
\alias{rf}
\title{Random Forest Model}
\description{wrapper}
\usage{rf(y = "load", data, for.each = "hour", cores = detectCores() - 1)}
\arguments{
  \item{y}{Character. Variable name.}
  \item{data}{data.frame}
  \item{for.each}{Character. Variable name.}
  \item{cores}{Parallel execution}
}
\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}

