% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sparse}
\alias{sparse}
\title{sparse}
\usage{
sparse(train.data, natoms, lambda, delta = 24)
}
\arguments{
\item{train.data}{an NxM matrix where each column is an hour by hour variable (load, temperature, etc.)}

\item{natoms}{normalization mode}

\item{lambda}{number of atoms in target dictionary}

\item{delta}{penalty term to be used}
}
\value{
a list with dictionaries
}
\description{
The predictive model is based on a "colinearity" argument between one day and
the other
Each traning sample consists of two consecutive days (Ltoday,Ltomorrow). A 
dictionary is then learnt to represent a set of training samples with that 
form. The dictionary will then be an ensemble of two-day atoms, the upper 
part corresponding to "todays", and the lower part to "tomorrows", 
[Dtod^T Dtom^T]^T
To predict tomorrow's load, w represent today's load Ltod in terms of Dtod 
and obtain a set of coefficients a
Whe then estimate Ltom as Dtom*a. 
The underlying hypothesis is that "a" would be essentially the same if we had asked to represent the whole vector [Ltod,Ltom]
}
