% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsb.R
\name{tsb}
\alias{tsb}
\title{tsb}
\usage{
tsb(yt, p = 1, d = 1, q = 1, P1 = 1, D1 = 0, Q1 = 1, P2 = 0,
  D2 = 1, Q2 = 1, s1 = 24, s2 = 168, k = 200)
}
\arguments{
\item{yt}{load data sequence. type numeric.}

\item{p}{AR order. type numeric.}

\item{d}{Degree of differencing. type numeric.}

\item{q}{MA order. type numeric.}

\item{P1}{AR order. type numeric.}

\item{D1}{Degree of differencing. type numeric.}

\item{Q1}{MA order. type numeric.}

\item{P2}{AR order. type numeric.}

\item{D2}{Degree of differencing. type numeric.}

\item{Q2}{MA order. type numeric.}

\item{s1}{Seasonal frequencies. type numeric.}

\item{s2}{Seasonal frequencies. type numeric.}

\item{k}{number of lags}
}
\value{
A list with ARIMA estimated model
\describe{
  \item{ar}{(reg=ar,est1=AR1,est2=AR2)}
  \item{ma}{(reg=ma,est1=MA1,est2=MA2)}
  \item{d}{(d,D1,D2)}
  \item{frec}{}
  \item{dif}{}
  \item{iter}{}
  \item{data}{load data}
}
}
\description{
Time series benchmark
}
\references{
{Brockwell, P.J.and Davis, R.A. 1991 \emph{Time Series: theory and methods} 2nd ed. Springer.}
}
\seealso{
\code{\link{predict.tsb}}
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
