% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavkerfun.R
\name{wavkerfun}
\alias{wavkerfun}
\title{Calibrate the KWF model}
\usage{
wavkerfun(obj, dist = 1, r, H, EPS = 1e-06)
}
\arguments{
\item{obj}{}

\item{dist}{}

\item{r}{}

\item{H}{}

\item{EPS}{}
}
\value{
fitted wavkerfun model
}
\description{
Using
}
