\name{werdist}
\alias{werdist}
\alias{addvalues}
\alias{adjust.noctave}
\alias{adjust.s0}
\alias{adjust.timeseries}
\alias{arealtest}
\alias{as.wt}
\alias{checkarealsiglevel}
\alias{checkweights}
\alias{coherence}
\alias{contrib}
\alias{createar}
\alias{createwavesurrogates}
\alias{createwgn}
\alias{criticalvaluesWCO}
\alias{criticalvaluesWSP}
\alias{CVkerfon}
\alias{cwt.ts}
\alias{DistWav}
\alias{foldKernel}
\alias{invmorlet}
\alias{jet.colors}
\alias{kernelArea}
\alias{kernelBitmap}
\alias{kernelRoot}
\alias{loss}
\alias{myimg}
\alias{plotat}
\alias{plotcoi}
\alias{plotcv}
\alias{plotmarks}
\alias{plotwt}
\alias{plot.wt}
\alias{prevkerfon}
\alias{rawWSP}
\alias{readmatrix}
\alias{readts}
\alias{rk}
\alias{scaleKernel}
\alias{scalematrix}
\alias{select.wkdata}
\alias{slide}
\alias{slope}
\alias{smCWT}
\alias{smooth.matrix}
\alias{smooth.time}
\alias{surrwave}
\alias{surrwave.character}
\alias{surrwave.matrix}
\alias{surrwave.ts}
\alias{surrwave.wt}
\alias{tobin}
\alias{toCWT}
\alias{toDWT}
\alias{vect2mat}
\alias{wavkerfun}
\alias{wco}
\alias{wcs}
\alias{werDist}
\alias{wgt}
\alias{wkdata}
\alias{writematrix}
\alias{wsp}
\alias{wtcolors}



%- Also NEED an '\alias' for EACH other topic documented here.
\title{todo wkf}
\description{todo wkf para estudiar y documentar}
\usage{werdist(data, nvoice = 4)}
\arguments{
  \item{data}{data.frame}
  \item{nvoice}{4}
}
\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}
