% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkdata.R
\name{wkdata}
\alias{wkdata}
\title{wkdata object constructor}
\usage{
wkdata(X, p, gr, J)
}
\arguments{
\item{X}{matrix of load curves (in columns)}

\item{p}{}

\item{gr}{groups in data (e.g. daytype)}

\item{J}{truncation level (max scale for the DWT)}
}
\value{
wkdata object
}
\description{
Constructs a wkdata object by taking the discrete wavelet transform DWT on 
the daily load curves.
}
