% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hong.R
\name{hong}
\alias{hong}
\title{hong}
\usage{
hong(data)
}
\arguments{
\item{data}{Data frame with columns:
\describe{
  \item{year}{year (...2012, 2013, 2014...)}
  \item{month}{month (01, 02, ..., 12)}
  \item{wday}{day of the week (Mon Thu Wed Tue Fri Sat Sun)}  
  \item{instant}{(01, 02, ..., 24) for hourly data or 
                 (01, 02, ..., 48) or half hourly data}  
  \item{temp}{temperature} 
  \item{load}{electricity loads}  
}}
}
\value{
An object of class \code{lm}
}
\description{
Fit Hong Vanilla Benchmark multiple regression model for electricity load
forecasting
}
\details{
The Hong Vanilla Model:
 \deqn{E(Load) = \beta_0 + \beta_1 Trend +\beta_2 Day \times Hour  + 
                 \beta_3 Month + \beta_4 Month \times Temp + 
                 \beta_5 Month  \times Temp^2  +\beta_6 Month  \times Temp^3 + 
                 \beta_7 Hour \times Temp + \beta_8 Hour \times Temp^2 +
                 \beta_9  Hour \times Temp^3}
}
\examples{
 1
}
\references{
Hong, T. and Wang, P. and Lee Willis, H. (2011) A naive multiple 
linear regression benchmark for short term load forecasting. In \emph{Power 
and Energy Society General Meeting, 2011 IEEE}, 1--6.

Hong, T. and Pinson, P.and Fan, S. (2014) Global energy forecasting 
competition 2012. \emph{International Journal of Forecasting}, 
\bold{30(2)}:357--363
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
