% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newvars.rf.R
\name{newvars.rf}
\alias{newvars.rf}
\title{newvars.rf}
\usage{
newvars.rf(data, type, t, ...)
}
\arguments{
\item{data}{Data frame with 'load' and 'temp' columns}

\item{type}{Character. 'load.lag', 'temp.lag', 'temp.sqcb', 'temp.desv', 'load.desv'.}

\item{t}{Lags}

\item{...}{...}
}
\value{
data.frame
}
\description{
Construct new variables for Random Forest Model
}
\details{
\describe{
  \item{load.lag}{lags of load}
  \item{temp.lag}{lags of temperature}
  \item{temp.sqcb}{temperature square and cubic}
  \item{ temp.desv}{deviation of temperature}
  \item{load.desv}{deviation of load}
}
}
\references{
{Hong, T. and Wang, P. and Lee Willis, H. 2011 A naive multiple linear regression benchmark for short term load forecasting. In \emph{Power and Energy Society General Meeting, 2011 IEEE}, 1--6.}

{Hong, T. and Pinson, P.and Fan, S. 2014 Global energy forecasting competition 2012. \emph{International Journal of Forecasting}, \bold{30(2)}:357--363}
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
