% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{performance}
\alias{performance}
\title{performance}
\usage{
performance(y, yhat, measure = "rmse")
}
\arguments{
\item{y}{vector of observed load}

\item{yhat}{vector of predicted load}

\item{measure}{\describe{
  \item{rmse} {Root Mean Square Error: \deqn{rmse(y,\hat{y})=\sqrt{\frac{1}{N}\sum\limits_{t=1}^N (y_t - \hat{y}_t)^2}}}
  \item{mape} {Mean Absolute Percentage Error: \deqn{mape(y,\hat{y})=\frac{1}{N}\sum\limits_{t=1}^N \frac{|y_t - \hat{y}_t|}{|y_t|}}}
  \item{mae} {Mean Absolute Error:\deqn{mae(y,\hat{y}) =\frac{1}{N}\sum\limits_{t=1}^N |y_t - \hat{y}_t|}}
  \item{pdad} {Promedio de Desviacion Absoluta de la Demanda: \deqn{pdad(y,\hat{y})=\frac{1}{N}\sum\limits_{t=1}^N \frac{|y_t - \hat{y}_t|}{\hat{y}}*100}}
}}
}
\value{
single value
}
\description{
Performance measure for a predicted set of values.
}
\examples{
performance(rep(0, 10), rnorm(10))
}
\references{
{  Hyndman, R.J. and Athanasopoulos, G. 2014 \emph{Forecasting: principles and practice}. OTexts.}
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
