% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.hong.R
\name{predict.hong}
\alias{predict.hong}
\title{predict.hong}
\usage{
\method{predict}{hong}(obj.hv, newdata = NULL)
}
\arguments{
\item{obj.hv}{Object of class inheriting from use of \code{hong} function.}

\item{new.data}{Data frame with predictors. If omitted, the fitted values are used.}
}
\value{
predicted values
}
\description{
Predict method for Hong Vanilla Model
}
\details{
\code{new.data} must contain the same predictors that data in \link{hong}.
Aditionally \code{new.data} must contain \code{trend} variable. This variable 
was computed like a sequence from 1 to number of rows of input data.frame in 
\link{hong} model function.
In forecasting new consecutive periods, \code{trend} variable in new.data 
must be computed like a sequence from number of rows of original data.
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
