\name{predict.rf}
\alias{predict.rf}
\title{Prediction in Random Forest}
\description{Predicted values for Random Forest Model}
\usage{predict.rf(listRF, newdata = NULL, for.each = "hour")}
\arguments{
  \item{listRF}{List of Random Forest Models}
  \item{newdata}{New data to predict}
  \item{for.each}{Character. Variable name of data frequency (i.e. 'hour')}
}
\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}

