\name{predict.ssm}
\alias{predict.ssm}
\title{Prediction in Univariate State Space Model}
\description{ VickyNando}
\usage{predict.ssm(ssm.obj, newdata, for.each = "hour")}
\arguments{
  \item{ssm.obj}{Estimated object from \code{ssm}}
  \item{newdata}{data }
  \item{for.each}{'hour'}
}
\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}

