% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tsb.R
\name{predict.tsb}
\alias{predict.tsb}
\title{predict.tsb}
\usage{
\method{predict}{tsb}(mod, steps = 24)
}
\arguments{
\item{mod}{An ARIMA model object fitting from \code{tsb}}

\item{steps}{Number of steps to forecast}
}
\value{
vector of predicted values
}
\description{
Predicted values for Time Series Benchmark Model
}
\seealso{
\link{tsb} \link{makeARIMA}
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
