\name{predict.tsb}
\alias{predict.tsb}
\title{Predict method for TSB Model}
\description{Predicted values for Time Series Benchmark Model}
\usage{predict.tsb(mod, steps = 24)}
\arguments{
  \item{mod}{An ARIMA model object fitting from \code{tsb}}
  \item{steps}{Number of steps to forecast}
}
\details{poner detalles}
\value{vector of predicted values}
\author{Jairo Cugliari, Andres Castrillejo, Fernando Massa}
\seealso{\link{tsb} \link{makeARIMA}}
