% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.wavkerfun.R
\name{predict.wavkerfun}
\alias{predict.wavkerfun}
\title{Title}
\usage{
\method{predict}{wavkerfun}(object, ..., newdata = NULL, sd)
}
\arguments{
\item{object}{}

\item{...}{}

\item{newdata}{}

\item{sd}{}
}
\description{
Title
}
