% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{rf}
\usage{
rf(y = "load", data, for.each = "hour", cores = parallel::detectCores() -
  1)
}
\arguments{
\item{y}{Character. Variable name.}

\item{data}{data.frame}

\item{for.each}{Character. Variable name.}

\item{cores}{Parallel execution}
}
\description{
Sets up random forest model for prediction
}
\author{
Jairo Cugliari, Andres Castrillejo, Fernando Massa, Ignacio Ramirez
}
