% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{rf}
\usage{
rf(y = "load", data, for.each = "hour", cores = parallel::detectCores() -
  1)
}
\arguments{
\item{y}{}

\item{data}{}

\item{for.each}{}

\item{cores}{}
}
\description{
Sets up random forest model for prediction
}
