% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{Title}
\usage{
ssm(formulaXa, formulaXf = NULL, data, ini = NULL, for.each = "hour",
  cores = parallel::detectCores() - 1)
}
\arguments{
\item{formulaXa}{}

\item{formulaXf}{}

\item{data}{}

\item{ini}{}

\item{for.each}{}

\item{cores}{}
}
\description{
Title
}
