% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsb.R
\name{tsb}
\alias{tsb}
\title{tsb}
\usage{
tsb(yt, p = 1, d = 1, q = 1, P1 = 1, D1 = 0, Q1 = 1, P2 = 0,
  D2 = 1, Q2 = 1, s1 = 24, s2 = 168, k = 200)
}
\arguments{
\item{yt}{}

\item{p}{}

\item{d}{}

\item{q}{}

\item{P1}{}

\item{D1}{}

\item{Q1}{}

\item{P2}{}

\item{D2}{}

\item{Q2}{}

\item{s1}{}

\item{s2}{}

\item{k}{}
}
\description{
Time series benchmark
}
